feat: Add permission checking screen
Introduces a new `PermissionsCheckActivity` that displays the status of all required app permissions in a recycler view. This screen allows users to grant any missing permissions directly.

The app now checks for missing permissions on resume and directs the user to this new screen, improving the permission handling flow.

Additionally, the message in the support dialog has been updated, and the logic for checking exact alarm permissions in the intro has been simplified.

Signed-off-by: Pranav Purwar <purwarpranav80@gmail.com>